backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
pydantic>=2.7.0
openai>=1.30.0
instructor>=1.3.0
langgraph>=0.1.0
langchain-core>=0.2.0
redis[hiredis]>=5.0.0
httpx>=0.27.0
pyyaml>=6.0.1
prometheus-client>=0.20.0
python-dotenv>=1.0.0
